chore: generate docs with sphinx
