+++ docs/index.rst
@@ -1,0 +1,20 @@
+Welcome to GraphQL-Server documentation!
+========================================
+
+Contents
+--------
+
+.. toctree::
+   :maxdepth: 2
+
+   intro
+   usage/index
+   modules/graphql-server
+
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
