+++ docs/modules/aiohttp.rst
@@ -1,0 +1,6 @@
+AIOHTTP
+=======
+
+.. currentmodule:: graphql_server.aiohttp
+
+.. automodule:: graphql_server.aiohttp
