+++ docs/modules/flask.rst
@@ -1,0 +1,6 @@
+FLASK
+=====
+
+.. currentmodule:: graphql_server.flask
+
+.. automodule:: graphql_server.flask
