+++ docs/modules/graphql-server.rst
@@ -1,0 +1,33 @@
+Reference
+=========
+
+.. currentmodule:: graphql_server
+
+.. autofunction:: run_http_query
+.. autofunction:: encode_execution_results
+.. autofunction:: format_execution_result
+.. autofunction:: json_encode
+.. autofunction:: load_json_body
+
+.. _data-structures:
+
+Data Structures
+---------------
+
+.. autoclass:: GraphQLParams
+.. autoclass:: GraphQLResponse
+.. autoclass:: ServerResponse
+.. autoclass:: HttpQueryError
+
+.. _sub-packages:
+
+Sub-Packages
+------------
+
+.. toctree::
+   :maxdepth: 1
+
+   aiohttp
+   flask
+   sanic
+   webob
