+++ docs/modules/sanic.rst
@@ -1,0 +1,6 @@
+SANIC
+=====
+
+.. currentmodule:: graphql_server.sanic
+
+.. automodule:: graphql_server.sanic
