+++ docs/modules/webob.rst
@@ -1,0 +1,6 @@
+WEBOB
+=====
+
+.. currentmodule:: graphql_server.webob
+
+.. automodule:: graphql_server.webob
