+++ docs/requirements.txt
@@ -1,0 +1,2 @@
+sphinx>=3.0.0,<4
+sphinx_rtd_theme>=0.4,<1
