+++ docs/usage/index.rst
@@ -1,0 +1,9 @@
+Usage
+=====
+
+GraphQL-Server provides the capability of integrating Graphql-Core features into
+any server integration.
+
+.. toctree::
+   :maxdepth: 2
+
