+++ tox.ini
@@ -52,3 +52,9 @@ basepython = python3.8
 deps = -e.[dev]
 commands =
     check-manifest -v
+
+[testenv:docs]
+basepython = python3.8
+deps = -e.[dev]
+commands =
+    sphinx-build -b html -EW docs docs/_build/html
